修复application名字不一致导致的bug；修复部分字段由Int改为double导致的bug

+++ src/main/java/cn/edu/ustb/sem/material/web/model/MaterialModel.java
@@ -86,13 +86,13 @@ public class MaterialModel {
 	public double getSingleNum() {
 		return singleNum;
 	}
-	public void setSingleNum(int singleNum) {
+	public void setSingleNum(double singleNum) {
 		this.singleNum = singleNum;
 	}
 	public double getBkNum() {
 		return bkNum;
 	}
-	public void setBkNum(int bkNum) {
+	public void setBkNum(double bkNum) {
 		this.bkNum = bkNum;
 	}
