+++ src/main/java/cn/edu/ustb/sem/order/web/model/OrderProcessModel.java
@@ -117,7 +117,7 @@ public class OrderProcessModel {
 		return measure;
 	}
 
-	public void setMeasure(int measure) {
+	public void setMeasure(double measure) {
 		this.measure = measure;
 	}
